Extended the range of dates that can be used
The old version was accurate all the way back to June 2021, this one is accurate all the way back to October 2019, which is a huge improvement and let's you get a lot more useful data. Not planning to go any further into the past as it's pretty much irrelevant(only interesting for historical and not practical reasons).
Also added 'delta' parameter which lets you change the default farm cycle from 7 days to anything you want. Comes with wrongly scaled data, but lets you get more resolution and see the trends in more detail. The scale can be manually adjusted by anyone who's going to use it, but I am not bothering because it is way beyond my usecase